test: ensure bpy mock import
