--- utils/audio_utils.py
@@ -1,7 +1,0 @@
-from blender_core import (
-    add_audio_strip,
-    split_audio_strip,
-    cut_audio_strip,
-    set_audio_volume,
-    extract_audio_from_video,
-)  # noqa: F401
