--- utils/clip_utils.py
@@ -1,1 +1,0 @@
-from blender_core import split_video_strip, cut_video_strip  # noqa: F401
