--- utils/transform_utils.py
@@ -1,1 +1,0 @@
-from blender_core import rotate_strip, translate_strip  # noqa: F401
